requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
PyJWT==1.7.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.16.0
twilio==6.60.0
urllib3==1.26.5

